move src/pages/ReactionPage/ to src/components/templates/ReactionPage/
