+++ src/components/templates/ReactionPage/ReactionPage.scss
@@ -1,4 +1,4 @@
-@import "scss/constants.scss";
+@import "scss/constants";
 
 .reaction-page-container {
   flex-direction: column;
